implemented restrictions by login_required

+++ basic/views.py
@@ -1,8 +1,8 @@
 from django.shortcuts import render, redirect
-from django.contrib import messages
-from django.urls import path
-from django.db.models import Q 
 from django.http import HttpResponse
+from django.contrib import messages
+from django.db.models import Q 
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate , login , logout
 from django.contrib.auth.models import User
 from .forms import RoomForm
@@ -17,6 +17,9 @@ from .models import Room , Topic
 # ]
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect("home")
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -64,6 +67,7 @@ def room(request , pk):
     context = {'room':room}
     return render(request,'basic/room.html', context)
 
+@login_required(login_url = 'login')
 def createRoom(request):
     form = RoomForm()
 
@@ -76,9 +80,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'basic/room_form.html', context)
 
+@login_required(login_url = 'login')
 def updateRoom(request , pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        return HttpResponse("Your not allowed here")
 
     if request.method == 'POST':
         form = RoomForm(request.POST , instance = room)
@@ -89,9 +97,13 @@ def updateRoom(request , pk):
     context = {'form' : form}
     return render(request, 'basic/room_form.html', context)
 
-
+@login_required(login_url = 'login')
 def deleteRoom(request , pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Your not allowed here")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
